validar formatos de formularios de pago

+++ js/pago.js
@@ -26,7 +26,7 @@ function mostrarDetallesCarro() {
 
         contenedorDetalles.appendChild(productoDiv);
 
-        total +=  producto.precioConIva * producto.cantidad;
+        total += producto.precioConIva * producto.cantidad;
     });
 
     document.getElementById('total-a-pagar').textContent = total.toFixed(2);
@@ -42,21 +42,67 @@ const FINALIZARSWAL = Swal.mixin({
     position: "center",
     showConfirmButton: true,
     showCancelButton: true,
-    
-  });
-const FINALIZAR= document.getElementById("finalizar");
-FINALIZAR.addEventListener("click",()=>{
-    FINALIZARSWAL.fire();
-    const OK= document.querySelector(".swal2-confirm");
 
-    OK.addEventListener("click",()=>{
-        window.location.href = "/index.html";
-        localStorage.setItem("CompraRealizada", "si");
-    })
-})
+});
+const REVISATUSDATOS = Swal.mixin({
+    title: "Datos incorrectos",
+    toast: true,
+    position: "center",
+    showConfirmButton: true,
+
+});
 //CERRAR SESION
 document.getElementById("cerrar-sesion").addEventListener("click", function CerrarSesion() {
     localStorage.removeItem("logCorrecto");
-  
+
     window.location.href = "/html/inicio.html";
-  });
+});
+
+
+
+//VALIDAR FORMULARIO ANTES DE ENVIAR
+
+document.getElementById('formulario-pago').addEventListener('submit', function (event) {
+    event.preventDefault(); // Evitar que se mande automaticamente
+
+    // llamar a los formularios
+    let nombre = document.getElementById('datos-nombre').value;
+    let direccion = document.getElementById('datos-direccion').value;
+    let telefono = document.getElementById('datos-telefono').value;
+    let tarjeta = document.getElementById("num-tarjeta").value;
+
+
+    // VALIDAR
+    var formularioEsValido = validarFormato(nombre, direccion, telefono, tarjeta);
+
+
+    // si es true muestra alerta para confirmar compra
+    if (formularioEsValido) {
+        const FINALIZAR = document.getElementById("finalizar");
+        FINALIZAR.addEventListener("click", () => {
+            FINALIZARSWAL.fire();
+            const OK = document.querySelector(".swal2-confirm");
+
+            OK.addEventListener("click", () => {
+                window.location.href = "/index.html";
+                localStorage.setItem("CompraRealizada", "si");
+            })
+        })
+
+    } else //si es false
+    {
+        const FINALIZAR = document.getElementById("finalizar");
+        FINALIZAR.addEventListener("click", () => {
+            REVISATUSDATOS.fire();
+        })
+    };
+});
+function validarFormato(nombre, direccion, telefono, tarjeta) {
+//revisar las condiciones
+    if (nombre.length < 4 || tarjeta.length != 16 || direccion.length < 4 || telefono.length < 5 ) {
+        return false; // devuelve false si no se cumple
+    } else {
+        return true;
+        // devuelve true
+    }
+};
